+++ lorry.cpp
@@ -6,7 +6,7 @@ Lorry::Lorry()
 
 }
 
-void Lorry::getData(float _cina, float _val, int _volume, QString _typ_oil, float _weight, int _year, bool _eur_1)
+void Lorry::getData(float _cina, float _val, float _volume, QString _typ_oil, float _weight, int _year, bool _eur_1)
 {
     Vehicle::getData(_cina,_val,_year);
     volume = _volume;
