+++ motocycle.h
@@ -5,12 +5,12 @@
 
 class Motocycle: public Vehicle
 {
-    int volume;
+    float volume;
     QString typ_oil;
     bool eur_1;
 public:
     Motocycle();
-    void getData(float _cina, float _val, int _volume, QString _typ_oil, bool _eur_1);
+    void getData(float _cina, float _val, float _volume, QString _typ_oil, bool _eur_1);
     void Calculate();
 };
